+++ client/src/components/Column/Column.scss
