+++ client/src/components/Kanban/index.js
@@ -1,5 +1,5 @@
 import "./Kanban.css";
-import List from "./basic/List";
+import List from "../basic/List";
 
 export default class Kanban extends List {
   constructor() {
