+++ client/src/components/basic/List/Li.js
@@ -1,0 +1,6 @@
+export default class Li extends Element {
+  constructor(content) {
+    super("li");
+    super.appendChild(content);
+  }
+}
